+++ src/user_handlers_test.go
@@ -19,7 +19,7 @@ func userTestSetup() {
 		http.Handle("/sona/v1/users", usrRouter)
 	}
 
-	userManager = RuntimeUserManager{make(map[int]*User), make(map[int]string), make(map[int][]string), make([]string, 0)}
+	userManager = RuntimeUserManager{make(map[int64]*User), make(map[int64]string), make(map[int64][]string), make([]string, 0)}
 	hookManager = HookManager{make([]WebHook, 0), make([]WebHook, 0), make([]WebHook, 0), make([]WebHook, 0), make([]WebHook, 0)}
 }
