+++ src/user_manager.go
@@ -2,11 +2,11 @@ package main
 
 type UserManager interface {
 	AddUser(user *AddUser) (bool, User)
-	GetUser(userId int) (User, bool)
-	UpdateUser(userId int, user *User) bool
-	RemoveUser(userId int) bool
+	GetUser(userId int64) (User, bool)
+	UpdateUser(userId int64, user *User) bool
+	RemoveUser(userId int64) bool
 	SetUserPassword(user User, password string)
-	SetPermissions(userId int, permissions []string) bool
+	SetPermissions(userId int64, permissions []string) bool
 	AuthenticateUser(user User, password string) (bool, TokenResponse)
 	ValidateUser(token string) bool
 }
